Get the x, y coordinates from a grid reference

Given an OSGB or OSNI grid reference string, get the x, y coordinates of the OSGB or OSNI grid for the bottom, left-hand corner of the grid square. The units parameter controls the units (metres m or kilometres km) in which the coordinates should be returned.

gridCoords(x, ...)

Arguments

grid_reference
A Great British or Irish grid reference character string with class gridref.

units
Should the returned coordinates be in metres (m) or kilometres (km).

Value

A list of class gridref with the following contents:

grid_reference: the original grid reference.

system: the grid reference projection system, either "OSGB" or "OSNI".

x: the easting coordinate in requested units.

y: the northing coordinate in requested units.

units: the requested units either metres (m) or kilometres (km).

precision: the precision of the original grid reference in metres

See also

Author

Stuart Ball.