Get 10km grid reference

This function returns the 10km grid reference for a higher precision grid reference. It uses the gridRef function in the archived rnbn package.

# S3 method for gridref
hectad(grid)

Arguments

grid_reference
character, British or Irish grid reference

Value

character, 10km grid reference

Details

It can convert either British or Irish grid references up to 10 figure (1m precision), including tetrads (2000m precision)

See also

Examples

if (FALSE) {
suppressPackageStartupMessages({
  library(store)
})

# add ten_km column
nbn_demonstration_dataset %>%
  janitor::clean_names() %>%
  dplyr::select(grid_reference) %>%
  dplyr::mutate(grid_reference = as_gridref(grid_reference)) %>%
  dplyr::rowwise() %>%
  dplyr::mutate(ten_km = hectad(grid_reference))
}