hectad_api

This function returns the 10km grid reference for a higher precision grid reference.

hectad_api(grid_reference, centre = FALSE)

Arguments

grid_reference
A Great British or Irish grid reference character string with class gridref.

Value

The grid reference at 10km resolution.

Details

It can check either British or Irish grid references up to 10 figure (1m precision), including tetrads (2000m precision).

REST API

A REST API endpoint is available once a Plumber router has been created and can be requested by adding _api to the end of the function.

The API returns "" instead of NULL so that it may be used with the dplyr::mutate function. Wrap the API call with dplyr::na_if to convert "" to NA.

suppressPackageStartupMessages({
  library(store)
  suppressWarnings({
    library(plumber)
  })
})
pr("../R/gridref_api.R") %>%
  pr_run(port = 8000)

See also

Examples

 suppressPackageStartupMessages({
   library(store)
   suppressWarnings({
     library(dplyr)
     library(janitor)
   })
 })

 grid_references <- nbn_demonstration_dataset %>%
   clean_names() %>%
   slice_tail(n = 15) %>%
   select(grid_reference) %>%
   mutate(grid_reference = as_gridref(grid_reference))

# add ten_km column
suppressPackageStartupMessages({
  suppressWarnings({
    library(dplyr)
  })
})

grid_references %>%
  rowwise() %>%
  mutate(ten_km = hectad(grid_reference))

#> # A tibble: 15 x 2
#> # Rowwise: 
#>    grid_reference ten_km
#>    <gridref>      <chr> 
#>  1 TL2384         TL28  
#>  2 TL232839       TL28  
#>  3 TL2385         TL28  
#>  4 TL231840       TL28  
#>  5 TL232839       TL28  
#>  6 TL23248346     TL28  
#>  7 TL23568493     TL28  
#>  8 TL23448505     TL28  
#>  9 TL23128324     TL28  
#> 10 TL234850       TL28  
#> 11 TL234846       TL28  
#> 12 TL232839       TL28  
#> 13 TL231841       TL28  
#> 14 TL226846       TL28  
#> 15 TL2384         TL28