reduce_image_size

This function reduces the size of an image, archiving the original image in a separate archive sub-directory

reduce_image_size(path_image, image_size = "50%")

Arguments

path_image
character, path name of image

image_size
percentage size reduction, default = 50%

Value

character, path name of image returned invisibly

See also

Examples

if (FALSE) {
suppressPackageStartupMessages({
  library(store)
  suppressWarnings({
    library(palmerpenguins)
    library(fs)
    library(dplyr)
    library(ggplot2)
    library(here)
    library(magick)
    library(future)
    library(furrr)
  })
})
# create output directory
i_am("example.Rmd")
if (!dir_exists("output")) {dir_create("output")}
if (!dir_exists("output/images")) {dir_create("output/images")}

# example taken from palmerpenguins example analysis of mass vs. flipper length
# https://allisonhorst.github.io/palmerpenguins/articles/examples.html

# data
mass_flipper_data <- penguins %>%
  select(species, flipper_length_mm, body_mass_g, )
mass_flipper_data

# graph
mass_flipper_graph <- ggplot(data = mass_flipper_data,
                             aes(x = flipper_length_mm,
                                 y = body_mass_g)) +
  geom_point(aes(color = species,
                 shape = species),
             size = 3,
             alpha = 0.8) +
  theme_minimal() +
  scale_color_manual(values = c("darkorange","purple","cyan4")) +
  labs(title = "Penguin size, Palmer Station LTER",
       subtitle = "Flipper length and body mass for Adelie, Chinstrap and Gentoo Penguins",
       x = "Flipper length (mm)",
       y = "Body mass (g)",
       color = "Penguin species",
       shape = "Penguin species") +
  theme(legend.position = c(0.2, 0.7),
        legend.background = element_rect(fill = "white", color = NA),
        plot.title.position = "plot",
        plot.caption = element_text(hjust = 0, face= "italic"),
        plot.caption.position = "plot")

# save graph
ggsave(filename = here("output", "images", "palmerpenguins_graph.png"),
       plot = mass_flipper_graph,
       type = "cairo-png",
       width = 8,
       height = 8,
       units = "in",
       dpi = 72)

# graph image
mass_flipper_graph <- image_read(here("output", "images", "palmerpenguins_graph.png"))
mass_flipper_graph

# graph information
image_info(mass_flipper_graph)

# image taken from Allison Horst's Art for teaching with palmerpengiuns
# Citation: Artwork by @allison_horst
lter_penguins <- image_read("https://allisonhorst.github.io/palmerpenguins/reference/figures/lter_penguins.png")
lter_penguins

# image information
image_info(lter_penguins)

# save image
image_write(image = lter_penguins,
            path = here("output", "images", "palmerpenguins_image.png"))

# batch reduce images
plan(multisession)
here("output/images") %>%
  dir_ls(., glob = "*.png") %>%
  future_map(reduce_image_size,
             image_size = "50%",
             .options = furrr_options(seed = TRUE),
             .progress = TRUE)

# reduced graph
mass_flipper_graph_reduce <- image_read(here("output", "images", "palmerpenguins_graph.png"))
mass_flipper_graph_reduce

# reduced graph information
image_info(mass_flipper_graph_reduce)

# reduced image
lter_penguins_reduce <- image_read(here("output", "images", "palmerpenguins_image.png"))
lter_penguins_reduce

# reduced image information
image_info(lter_penguins_reduce)
}